rename extend to extendKey
