+++ packages/unlock-js/src/PublicLock/v10/extendKey.js
